check for and enable validation layers
because:
- validation layers really help for catching
  vulkan mistakes
